feat:finalize api routes

+++ server/src/routes/memories.ts
@@ -3,8 +3,18 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get("/memories", async () => {
+  app.addHook("preHandler", async (req) => {
+    await req.jwtVerify();
+  });
+
+  app.get("/memories", async (req) => {
+    await req.jwtVerify();
+
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
+
       orderBy: {
         createdAt: "asc",
       },
@@ -19,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (req) => {
+  app.get("/memories/:id", async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -31,6 +41,10 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && memory.userId !== req.user.sub) {
+      return reply.status(401).send();
+    }
 
     return memory;
   });
@@ -49,14 +63,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "debd9994-aaad-4b15-9ea1-46b3c26174a4",
+        userId: req.user.sub,
       },
     });
 
     return memory;
   });
 
-  app.put("/memories/:id", async (req) => {
+  app.put("/memories/:id", async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -71,7 +85,17 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, isPublic, coverUrl } = bodySchema.parse(req.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== req.user.sub) {
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -85,12 +109,22 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory;
   });
 
-  app.delete("/memories/:id", async (req) => {
+  app.delete("/memories/:id", async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(req.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== req.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
